noe can greier

+++ node1/can.h
@@ -1,6 +1,6 @@
 #include "cancont.h"
 typedef struct {
-    uint32_t id; 
+    uint16_t id; 
     uint8_t dlc; //data length
     uint8_t data[8];
 } canMessage;
